gtk: split expression objects into multiple files
otherwise rustdoc-stripper strugles to add the docs to them
